base: clean up remaining compiler warnings across the tree
Same categories as the earlier kernel-only pass, applied to the rest
of base/:

- Catch ConnError/Error/ParsingError/TGDriveError by reference instead
  of by value across zelio.cpp, nut.cpp, script.cpp, tgdrive.cpp and
  d50.cpp, avoiding exception slicing (-Wcatch-value=).
- Null-terminate centrald.cpp's hostname after strncpy(), matching the
  device.cpp fix from the previous commit (-Wstringop-truncation).
- Cast size_t/size-typed arguments passed to printf '*'/'.*' width and
  precision specifiers in the ncurses monitor windows, and an int cast
  for a size_t compared against an int elsewhere
  (-Wformat=, -Wsign-compare).
- Widen a couple of command buffers (nstep.cpp, dut1.cpp) and use
  sizeof(buf) instead of a hand-counted literal (tellx200.cpp) so the
  compiler can prove snprintf can't truncate (-Wformat-truncation=).
- Un-hide two more base-class virtuals via 'using' declarations, only
  where doing so can't introduce an ambiguous call: NWindow::draw()
  in NSubmenu (nmenu.h) and Telescope::sky2counts() in Fork (fork.h).
  Neither hidden overload has a signature that a caller of the other
  could accidentally match, unlike ConnTCP::init(bool = true) which
  was left alone in the previous commit for exactly that reason.
- Initialize r0/n in Dummy::generateData() and totPix in
  Camera::sendReadoutData(): each is set inside a switch/if with no
  default/else, so an unmatched case previously left them
  uninitialized (-Wmaybe-uninitialized). r0 in particular is a real
  bug, not just a compiler nit: with genType==6 (artificial stellar
  field) and no focuser (getFocPos()==-1), r0 was read uninitialized
  in the per-pixel Moffat profile calculation.

Left three categories alone, same reasoning as before (not a
mechanical fix, needs a design decision or broader audit):
- device.cpp/daemon.cpp's createValue() reinterpret-cast pattern
  (-Wstrict-aliasing).
- Connection::init() hidden by ConnTCP::init(bool = true)
  (-Woverloaded-virtual=) - the default argument makes a using-
  declaration ambiguous at every zero-arg call site.
- camd.h's updateStatistics<t>(): modeCount is sized as
  1 << (sizeof(t)*8), which is undefined behaviour for 8-byte pixel
  types (double, int64_t) and a ~16GB allocation for 4-byte types
  (uint32_t) - the per-pixel mode-value histogram only really makes
  sense for 8/16-bit data. Needs a product decision (skip mode stats
  for wide types? clamp?), not a one-line fix.
- test_config.cpp/test_libnova_cpp.cpp's unused 'ret'/'hrz': both are
  only referenced inside assert(), which compiles to nothing under
  -DNDEBUG (this tree builds RelWithDebInfo by default) - the real
  issue is that these tests silently pass-by-doing-nothing in that
  build type, not the unused variable itself.

+++ base/monitor/src/ndevicewindow.cpp
@@ -67,7 +67,7 @@ void NDeviceWindow::printState ()
 
 void NDeviceWindow::printValue (const char *name, const char *value, bool writable)
 {
-	wprintw (getWriteWindow (), "%c %-*s %30.*s\n", ((writable) ? 'W' : ' '), valueBegins, name, getScrollWidth () - valueBegins - 5, value);
+	wprintw (getWriteWindow (), "%c %-*s %30.*s\n", ((writable) ? 'W' : ' '), (int) valueBegins, name, (int) (getScrollWidth () - valueBegins - 5), value);
 }
 
 void NDeviceWindow::printValue (rts2core::Value * value)
@@ -91,7 +91,7 @@ void NDeviceWindow::printValue (rts2core::Value * value)
 	// ultra special handling of SCRIPT value
 	if (value->getValueDisplayType () == RTS2_DT_SCRIPT)
 	{
-		wprintw (getWriteWindow (), "  %-*s ", valueBegins, value->getName ().c_str ());
+		wprintw (getWriteWindow (), "  %-*s ", (int) valueBegins, value->getName ().c_str ());
 		wcolor_set (getWriteWindow (), CLR_DEFAULT, NULL);
 		const char *valStart = value->getValue ();
 		if (!valStart)
@@ -182,7 +182,7 @@ void NDeviceWindow::printValue (rts2core::Value * value)
 			}
 			break;
 		case RTS2_VALUE_SELECTION:
-			wprintw (getWriteWindow (), "%c %-*s %5i %24.*s\n", value->isWritable () ? 'W' : ' ', valueBegins, value->getName ().c_str (), value->getValueInteger (), getScrollWidth () - valueBegins - 10, ((rts2core::ValueSelection *) value)->getSelName ());
+			wprintw (getWriteWindow (), "%c %-*s %5i %24.*s\n", value->isWritable () ? 'W' : ' ', (int) valueBegins, value->getName ().c_str (), value->getValueInteger (), (int) (getScrollWidth () - valueBegins - 10), ((rts2core::ValueSelection *) value)->getSelName ());
 			break;
 		default:
 			printValue (value->getName ().c_str (), getDisplayValue (value).c_str (), value->isWritable ());
